create index for student profile hooks
