+++ frontend/src/components/StudentProfile/cards/BonusesCard.tsx
@@ -1,4 +1,4 @@
-import { Bonus } from "../../../hooks/StudentProfile/useStudentProfileData/useBonusesCardData";
+import { Bonus } from "../../../hooks/StudentProfile";
 import { EMPTY_FIELD_STRING } from "../../../utils/constants";
 import { Styles } from "../../../utils/Styles";
 import { AwardWithTooltip } from "../../images/AwardWithTooltip";
