+++ frontend/src/components/StudentProfile/table/StudentTableWithFilters.tsx
@@ -1,7 +1,7 @@
 import { Styles } from "../../../utils/Styles";
 import FilterMenu from "./FilterMenu";
 import { useState } from "react";
-import { Points } from "../../../hooks/StudentProfile/useStudentProfileData/useStudentData";
+import { Points } from "../../../hooks/StudentProfile";
 import { FilterItem } from "../../Groups/FilterBar/FilterOptionsSection";
 import { StudentTable } from "./StudentTable";
