+++ frontend/src/components/StudentProfile/table/cellContent/AwardsCellContent.tsx
@@ -1,4 +1,4 @@
-import { Points } from "../../../../hooks/StudentProfile/useStudentProfileData/useStudentData";
+import { Points } from "../../../../hooks/StudentProfile";
 import { EMPTY_FIELD_STRING } from "../../../../utils/constants";
 import { Styles } from "../../../../utils/Styles";
 import { AwardWithTooltip } from "../../../images/AwardWithTooltip";
