+++ frontend/src/components/StudentProfile/table/cellContent/DateCellContent.tsx
@@ -1,4 +1,4 @@
-import { Points } from "../../../../hooks/StudentProfile/useStudentProfileData/useStudentData";
+import { Points } from "../../../../hooks/StudentProfile";
 import { dateOptions } from "../../../../utils/constants";
 
 type DateCellContentProps = {
