+++ frontend/src/components/images/AwardImage.tsx
@@ -3,6 +3,7 @@ import { Image } from "./Image";
 type AwardImageProps = {
   id: string | undefined;
   size: AwardImageSize;
+  disabled?: boolean;
 };
 
 const sizeMap: Record<AwardImageSize, number> = {
@@ -12,6 +13,6 @@ const sizeMap: Record<AwardImageSize, number> = {
 
 export type AwardImageSize = "s" | "m";
 
-export const AwardImage = ({ id, size }: AwardImageProps) => {
-  return <Image id={id} size={sizeMap[size]} />;
+export const AwardImage = ({ id, size, disabled = false }: AwardImageProps) => {
+  return <Image id={id} size={sizeMap[size]} disabled={disabled} />;
 };
