+++ frontend/src/components/images/Image.tsx
@@ -14,7 +14,7 @@ type ImageProps = {
   disabled: boolean;
 };
 
-export const Image = ({ size, id, disabled: disabled }: ImageProps) => {
+export const Image = ({ size, id, disabled }: ImageProps) => {
   return (
     <div style={{ width: size, height: size, opacity: disabled ? 0.6 : 1 }}>
       <img src={`${FILES_URL}${id}`} alt={`img id ${id}`} style={styles.img} />
