+++ frontend/src/hooks/StudentProfile/index.ts
@@ -1,0 +1,4 @@
+export * from "./types";
+export * from "./useLevelsData";
+export * from "./useStudentProfileData";
+export * from "./useTeacherActions";
