+++ frontend/src/hooks/StudentProfile/useStudentProfileData/index.ts
@@ -1,0 +1,3 @@
+export * from "./useBonusesCardData";
+export * from "./useStudentData";
+export * from "./useStudentProfileData";
