+++ frontend/src/screens/StudentProfile/TeacherStudentProfile.tsx
@@ -2,13 +2,13 @@ import { Styles } from "../../utils/Styles";
 import { useParams } from "react-router-dom";
 import { PointsForm } from "../../components/StudentProfile/PointsForm/PointsForm";
 import { useUser } from "../../hooks/common/useUser";
-import { useStudentProfileData } from "../../hooks/StudentProfile/useStudentProfileData/useStudentProfileData";
+import { useStudentProfileData } from "../../hooks/StudentProfile";
 import { SideBar } from "../../components/StudentProfile/SideBar";
 import { useFormCategories } from "../../hooks/common/useFormCategories";
 import { Dialog } from "@mui/material";
 import { StudentTableWithFilters } from "../../components/StudentProfile/table/StudentTableWithFilters";
 import { Button } from "../../components/Button";
-import { useTeacherActions } from "../../hooks/StudentProfile/useTeacherActions";
+import { useTeacherActions } from "../../hooks/StudentProfile";
 import { useEditionSelection } from "../../hooks/common/useEditionSelection";
 import { Roles } from "../../router/paths";
 import { isEditionActive } from "../../utils/utils";
